+++ src/test/kotlin/snc/openchargingnetwork/node/integration/utils/Helpers.kt
@@ -36,3 +36,5 @@ fun SignableHeaders.toMap(tokenC: String, signature: String): MutableMap<String,
     toPartyId?.let { map["OCPI-To-Party-Id"] = it }
     return map
 }
+
+